Scripts/requirements.txt
pyyaml>=6.0.1
python-frontmatter>=1.0.0
pathlib>=1.0.1
networkx>=3.1
scikit-learn>=1.3.0
numpy>=1.24.0
